+++ common/src/main/java/mca/entity/interaction/InteractionPredicate.java
@@ -1,5 +1,7 @@
 package mca.entity.interaction;
 
+import java.util.LinkedList;
+import java.util.List;
 import java.util.Map;
 
 import mca.entity.interaction.gifts.GiftPredicate;
@@ -21,6 +23,7 @@ public class InteractionPredicate {
 
         @Nullable
         GiftPredicate.Condition condition = null;
+        List<String> conditionKeys = new LinkedList<>();
 
         for (Map.Entry<String, JsonElement> entry : json.entrySet()) {
             if ("chance".equals(entry.getKey())) {
@@ -29,6 +32,7 @@ public class InteractionPredicate {
                 hearts = JsonHelper.asInt(entry.getValue(), entry.getKey());
             } else if (CONDITION_TYPES.containsKey(entry.getKey())) {
                 GiftPredicate.Condition parsed = CONDITION_TYPES.get(entry.getKey()).parse(entry.getValue());
+                conditionKeys.add(entry.getKey());
                 if (condition == null) {
                     condition = parsed;
                 } else {
@@ -37,7 +41,7 @@ public class InteractionPredicate {
             }
         }
 
-        return new InteractionPredicate(chance, hearts, condition);
+        return new InteractionPredicate(chance, hearts, condition, conditionKeys);
     }
 
     private final float chance;
@@ -45,11 +49,13 @@ public class InteractionPredicate {
 
     @Nullable
     private final GiftPredicate.Condition condition;
+    List<String> conditionKeys;
 
-    public InteractionPredicate(float chance, int hearts, @Nullable GiftPredicate.Condition condition) {
+    public InteractionPredicate(float chance, int hearts, @Nullable GiftPredicate.Condition condition, List<String> conditionKeys) {
         this.chance = chance;
         this.hearts = hearts;
         this.condition = condition;
+        this.conditionKeys = conditionKeys;
     }
 
     public boolean test(VillagerEntityMCA villager) {
@@ -62,5 +68,9 @@ public class InteractionPredicate {
 
     public int getHearts() {
         return hearts;
+    }
+
+    public List<String> getConditionKeys() {
+        return conditionKeys;
     }
 }